buncha changes to work better with GitHub Actions, cover more code, etc

+++ tests/runtests.sh
@@ -2,33 +2,39 @@
 
 # Run tests. Output is compared with expected.out
 
-# If TEST_CRED is set, server, user and password are grepped out of it
-# and added to test.cfg. In that case, it is expected that the
-# referenced IMAP account will have a test-polly folder/label. This is
-# only used to exercise the read loop and increase coverage, so it's
-# not strctly necessary.
-
 export TMPDIR=$(mktemp -d)
-# trap "rm -rf ${TMPDIR}" EXIT
+trap "rm -rf ${TMPDIR}" EXIT
 
-if [ "x$1" = "x--generate" ] ; then
-    OUT=/dev/stdout
-    GENONLY=1
+OUT=${TMPDIR}/actual.out
+DOCOV=0
+
+while getopts 'c' OPTION; do
+    case "$OPTION" in
+        c)
+            DOCOV=1
+            ;;
+    esac
+done
+shift "$(($OPTIND -1))"
+
+if [ "x$DOCOV" = "x1" ] ; then
+    PYTHON='coverage run --concurrency=thread --append'
 else
-    OUT=${TMPDIR}/actual.out
-    GENONLY=0
+    PYTHON='python'
 fi
 
-coverage erase
-rm -rf htmlcov
+if [ "x$DOCOV" = "x1" ] ; then
+    coverage erase
+    rm -rf htmlcov
+fi
 for f in $(ls tests/cfgs/test-*.cfg) ; do
     echo "* $f"
-    coverage run --append src/polly.py -g 5 -c ${f}
+    ${PYTHON} src/polly.py -g 5 -c ${f}
     echo ""
 done > ${OUT}
 
-if [ "x$GENONLY" = "x0" ] ; then
-    echo '[Polly]
+cat <<EOF > ${TMPDIR}/test.cfg
+[Polly]
 verbose = trace
 folder = test-polly
 threshold = 0.25
@@ -39,31 +45,47 @@ upper = False
 maxchars = 6
 length = 4
 logfile = /dev/null
-picklefile = '${TMPDIR}'/polly.pkl
+picklefile = ${TMPDIR}/polly.pkl
 unittests = True
-' > ${TMPDIR}/test.cfg
+EOF
 
-    if [ "x$TEST_CRED" != "x" ] ; then
-        egrep '^ *(server|user|password) *=' $TEST_CRED >> ${TMPDIR}/test.cfg
-    fi
-    echo flippant > ${TMPDIR}/extra.words
-    echo 'bad djm efore etty fter ginal ilt ong tle
-    add '${TMPDIR}'/extra.words 1
-    good polly.good
-    dict american-english-large.txt
-    option logfile /tmp/trash
-    option digits false
-    option length 7
-    option editing-mode vi
-    read
-    sleep 5
-    sleep 5
-    sleep 5
-    sleep 5
-    sleep 5
-    ' | coverage run --concurrency=thread --append src/polly.py -n \
-                 -c ${TMPDIR}/test.cfg
-
-    coverage html
-    diff -u tests/output/expected.out ${OUT}
+if [ "x$TEST_CRED" != "x" ] ; then
+    egrep '^ *(server|user|password) *=' $TEST_CRED >> ${TMPDIR}/test.cfg
 fi
+
+echo flippant > ${TMPDIR}/extra.words
+cat <<EOF > ${TMPDIR}/test.cmds
+option logfile /tmp/trash
+option logfile ${TMPDIR}/test-run-bz2.log.bz2
+option logfile ${TMPDIR}/test-run-gz.log.gz
+option logfile /tmp/trash
+bad djm efore etty fter ginal ilt ong tle
+add ${TMPDIR}/extra.words 1
+good polly.good
+rebuild
+dict american-english-large.txt
+option digits false
+option length 7
+option editing-mode vi
+option verbose info
+password
+read
+sleep 5
+option verbose trace
+sleep 5
+option verbose debug
+sleep 5
+option verbose warning
+sleep 5
+option verbose trace
+sleep 5
+option verbose info
+stat
+EOF
+
+${PYTHON} src/polly.py -n -c ${TMPDIR}/test.cfg < ${TMPDIR}/test.cmds
+
+if [ "x$DOCOV" = "x1" ] ; then
+    coverage html
+fi
+diff -u tests/output/expected.out ${OUT}
